Revert "feat: refactor TableLoadingBody to TableLoadingRows"
This reverts commit 2483a5b3e312f0beb1ffa0db9f8f3ee9e89276c2.
