--- frontend/src/modeldirectory/components/tableLoadingRows/TableLoadingRows.test.tsx
@@ -1,28 +1,0 @@
-import { Table } from "@mui/material";
-import {render, screen} from "@testing-library/react";
-import TableLoadingRows, {DATA_TEST_ID} from "./TableLoadingRows";
-import TableBody from "@mui/material/TableBody";
-
-describe("TableLoadingRows", () => {
-  test("should render correctly", () => {
-    // GIVEN n numberOfCOls
-    const givenNumberOfCols = 5;
-    // AND n numberOfRows
-    const givenNumberOfRows = 10;
-    
-    // WHEN the TableLoadingRows is rendered with the given numberOfCols and numberOfRows
-    render(<Table>
-      <TableBody>
-        <TableLoadingRows numberOfCols={givenNumberOfCols} numberOfRows={givenNumberOfRows} />
-      </TableBody>
-    </Table>)
-    
-    // THEN expect the given number of rows to be shown
-    const skeletonRowElements = screen.getAllByTestId(DATA_TEST_ID.SKELETON_ROW);
-    expect(skeletonRowElements).toHaveLength(givenNumberOfRows);
-    // AND the given number of columns to be shown
-    const skeletonCellElements = screen.getAllByTestId(DATA_TEST_ID.SKELETON_CELL);
-    const availableNumberOfCols = skeletonCellElements.length / skeletonRowElements.length
-    expect(availableNumberOfCols).toBe(givenNumberOfCols);
-  })
-})
